Fixed incorrect information about setting minor and patch version numbers to 0.
The current Slovak translation says: "a minor version change can contain patch level changes, but in that case, the patch version number has to be set to 0" while the English specification says: "the patch version number must be set to 0 ALWAYS when a minor version number is changed". The same applies for major version change.