Add hover anchor links for spec headers and sections
This adds anchors for each heading by changing the markdown renderer to
Kramdown. Unfortunately, Kramdown doesn't have an option to do the same
for list items. Instead, we do a tiny bit of JavaScript to add IDs to
each list item element in the specification section.

Then we use some JavaScript to insert a link with an embedded base64
background image that is displayed on hover for each element that has an
ID.

Fixes #27
Closes #128